add: auth signin & signup wip
