+++ backend/app/controllers/api/v1/auth/sessions_controller.rb
@@ -1,6 +1,5 @@
 class Api::V1::Auth::SessionsController < ApplicationController
     def index
-        byebug
         if current_api_v1_user
             render json: { is_login: true, data: current_api_v1_user }
         else
