+++ frontend/app/src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css';
-import React from 'react';
+import React, { useState, useEffect, createContext } from 'react';
 import Home from './Home';
 import { Routes, Route } from "react-router-dom";
 import LocationPostDetail from './LocationPostDetail';
@@ -12,25 +12,72 @@ import PostUpdate from './PostUpdate'
 import SignUp from './SignUp';
 import SignIn from './SignIn';
 import UserProfile from './UserProfile';
+import { User } from './interfaces';
+import { getCurrentUser } from './lib/api/auth';
+
+export const AuthContext = createContext({} as {
+  // loading: boolean
+  // setLoading: React.Dispatch<React.SetStateAction<boolean>>
+  isSignedIn: boolean
+  setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>
+  currentUser: User | undefined
+  setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>
+})
+
 
 function App() {
+  const [isSignedIn, setIsSignedIn] = useState<boolean>(false)
+  const [currentUser, setCurrentUser] = useState<User | undefined>()
+
+
+  const handleGetUserCurrent = async() => {
+    console.log("#####access-token:"+localStorage.getItem('access-token'))
+    console.log("#####client:"+localStorage.getItem('cliet'))
+    console.log("#####uid:"+localStorage.getItem('uid'))
+  
+    try {
+        const response = await getCurrentUser()
+        if (response?.data.is_login === true) {
+            console.log(response.data);
+            setIsSignedIn(true)
+            setCurrentUser(response?.data.data)
+            console.log("current user")
+        } else {
+            localStorage.clear()
+            console.log("not current user")
+        }
+    } catch(err) {
+        console.log(err)
+    }
+  }
+
+  useEffect(() => {
+    handleGetUserCurrent()
+  }, [setCurrentUser])
+
+
+// TODO: Loadingと場合分け
+
+
 
   return (
     <div className="App">
-      <Routes>
-        <Route path="/" element={ <Home /> } />
+      <AuthContext.Provider value={{ isSignedIn, setIsSignedIn, currentUser, setCurrentUser }}>
+        <Routes>
+          <Route path="PostCreate" element={ <PostCreate /> }/>
+          <Route path="PostUpdate" element={ <PostUpdate /> }/>
+          <Route path="UserProfile" element= { <UserProfile /> }></Route>
+          <Route path="/" element={ <Home /> } />
         <Route path="/LocationPostDeatil" element={ <LocationPostDetail /> } />
         <Route path="/LocationPosts" element={ <LocationPosts /> }>
           <Route index element={ <LocationPostsIndex /> }/>
           <Route path=":locationPostId" element={ <LocationPost /> } />
         </Route>
-        <Route path="PostCreate" element={ <PostCreate /> }/>
-        <Route path="PostUpdate" element={ <PostUpdate /> }/>
         <Route path="SignUp" element= { <SignUp /> }></Route>
         <Route path="SignIn" element= { <SignIn /> }></Route>
-        <Route path="UserProfile" element= { <UserProfile /> }></Route>
         <Route path="*" element={ <NotFound /> } />
-      </Routes>
+        </Routes>
+      </AuthContext.Provider> 
     </div>
   );
 }
