+++ frontend/app/src/SignIn.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { useEffect, useState } from 'react';
+import { useEffect, useState, useContext } from 'react';
 
 import { Link, useNavigate } from "react-router-dom";
 import Header from './components/Header';
 import { signIn } from './lib/api/auth';
+import { AuthContext } from './App';
 
 interface SignInParams {
     email: string,
@@ -16,8 +17,9 @@ const SignIn: React.FC = () => {
         password: '',
     }
     const navigation = useNavigate()
-
+    const { setIsSignedIn, setCurrentUser } = useContext(AuthContext)
     const [SignInParams, setSignInParams] = useState<SignInParams>(blankParams)
+
 
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         setSignInParams({
@@ -39,6 +41,9 @@ const SignIn: React.FC = () => {
                 localStorage.setItem('access-token', response.headers['access-token']);
                 localStorage.setItem('client', response.headers['client']);
                 localStorage.setItem('uid', response.headers['uid']);
+
+                setIsSignedIn(true)
+                setCurrentUser(response.data.data)
                 navigation("/UserProfile");
             }
         } catch (error) {
