+++ frontend/app/src/interfaces/index.tsx
@@ -9,3 +9,16 @@ export interface SignInParams {
     email: string,
     password: string,
 }
+
+export interface User {
+    id: number
+    uid: string
+    provider: string
+    email: string
+    name: string
+    nickname?: string
+    image?: string
+    allowPasswordChange: boolean
+    created_at: Date
+    updated_at: Date
+}
